Change prettier configuration to allow 100 characters per line

+++ app/workers/clean_up.rb
@@ -12,18 +12,12 @@ class CleanUp
 
   # Remove users that have never logged in and created their account more than two years ago
   def remove_unused_accounts
-    User
-      .where(current_sign_in_at: nil)
-      .where("created_at < ?", 2.years.ago)
-      .destroy_all
+    User.where(current_sign_in_at: nil).where("created_at < ?", 2.years.ago).destroy_all
   end
 
   # Remove users that haven't confirmed their email in 4 weeks
   def remove_unconfirmed_accounts
-    User
-      .where(confirmed_at: nil)
-      .where("created_at < ?", 4.weeks.ago)
-      .destroy_all
+    User.where(confirmed_at: nil).where("created_at < ?", 4.weeks.ago).destroy_all
   end
 
   def anonymize_sign_up_ip
